+++ vnext/ReactUWP/Pch/pch.cpp
@@ -1,7 +1,0 @@
-﻿// Copyright (c) Microsoft Corporation.
-// Licensed under the MIT License.
-
-#include "pch.h"
-
-// Please don't add any file in the PCH.cpp. the file is not used in office
-// build.