+++ vnext/CHANGELOG.md
@@ -1,8 +1,27 @@
 # Change Log - react-native-windows
 
-This log was last generated on Thu, 25 Jun 2020 14:32:50 GMT and should not be manually modified.
+This log was last generated on Fri, 26 Jun 2020 23:59:27 GMT and should not be manually modified.
 
 <!-- Start content -->
+
+## 0.0.0-canary.104
+
+Fri, 26 Jun 2020 23:59:27 GMT
+
+### Changes
+
+- Upgrade React Native to 3/22 Nightly Build (ngerlem@microsoft.com)
+
+## 0.0.0-canary.103
+
+Fri, 26 Jun 2020 20:17:03 GMT
+
+### Changes
+
+- Add a null check on reload callback (acoates@microsoft.com)
+- Unify storage for sourceBundleHost/Port and debugHost (acoates@microsoft.com)
+- Remove ReactUWP project folder (vmorozov@microsoft.com)
+- Fix empty host getting passed to HMR client when SourceBundleHost not specified (acoates@microsoft.com)
 
 ## 0.0.0-canary.102